# Backend/openapi.yml
openapi: 3.0.0
info:
  title: Beacon API
  version: 1.0.0
  description: |
    Backend API for Beacon.

    Most `/api/*` endpoints require a JWT in `Authorization: Bearer <token>`.
    Request validation is enforced via `express-openapi-validator`.
servers:
  - url: http://localhost:3000
    description: Local development server
...
paths:
...
  /api/pins:
    get:
      tags: [Pins]
      summary: Get all pins
      description: |
        Retrieves all pins from the database. Supports optional distance sort via
        `sort=distance&lat=&lng=` (Near Me priority). When authenticated, includes
        the requesting user's status per pin (`visited` or `wishlist`).
      operationId: getAllPins
      parameters:
        - in: query
          name: sort
          required: false
          schema:
            type: string
            enum: [recent, distance]
          description: Sort mode. `distance` requires lat/lng query params.
        - in: query
          name: lat
          required: false
          schema:
            type: number
            minimum: -90
            maximum: 90
          description: Latitude of origin for distance sort.
        - in: query
          name: lng
          required: false
          schema:
            type: number
            minimum: -180
            maximum: 180
          description: Longitude of origin for distance sort.
      responses:
        200:
          description: Successful response with array of pins
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Pin'
        401:
          description: Unauthorized - invalid or missing token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    
    post:
      tags: [Pins]
      summary: Create a new pin
      description: Creates a new pin with the provided data
      operationId: createPin
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - latitude
                - longitude
              properties:
                latitude:
                  type: number
                  minimum: -90
                  maximum: 90
                  description: Latitude coordinate of the pin
                  example: 40.7128
                longitude:
                  type: number
                  minimum: -180
                  maximum: 180
                  description: Longitude coordinate of the pin
                  example: -74.0060
                title:
                  type: string
                  maxLength: 200
                  description: Title of the pin
                  example: "Golden Gate Overlook"
                description:
                  type: string
                  maxLength: 500
                  description: Description/message content of the pin
                  nullable: true
                image:
                  type: string
                  maxLength: 2000
                  description: URL of the image
                  nullable: true
                address:
                  type: string
                  maxLength: 200
                  description: Address of the pin location
                  nullable: true
                tags:
                  oneOf:
                    - type: string
                    - type: array
                      items:
                        type: string
                  description: Tags for the pin

      responses:
        201:
          description: Pin created successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: integer
                    description: ID of the newly created pin
        400:
          description: Bad request - missing required fields
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        401:
          description: Unauthorized - invalid or missing token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        500:
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /api/pins/trending:
    get:
      tags: [Pins]
      summary: Get trending pins
      description: |
        Returns up to 20 pins ordered by a trending score of
        `likes + 3 * max(0, 1 - ageDays / days)`. Recency score decays linearly
        and pins outside the window receive a recency contribution of 0.
      operationId: getTrendingPins
      parameters:
        - in: query
          name: days
          required: false
          schema:
            type: integer
            minimum: 1
            maximum: 365
            default: 7
          description: Number of days the linear recency decay is stretched over.
      responses:
        200:
          description: Successful response with trending pins
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Pin'
        401:
          description: Unauthorized - invalid or missing token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
  /api/pins/user:
    get:
      tags: [Pins]
      summary: Get all pins created by the logged-in user
      description: Retrieves all pins created by the authenticated user
      operationId: getUserPins
      responses:
        200:
          description: Successful response with array of user's pins
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Pin'
        401:
          description: Unauthorized - invalid or missing token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /api/pins/nearby:
    post:
      tags: [Pins]
      summary: Get pins near a coordinate
      description: Returns all pins within 10 km of the given latitude/longitude, sorted by distance (closest first), using the Haversine formula for accurate great-circle distance.
      operationId: getPinsNearCoordinate
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - latitude
                - longitude
              properties:
                latitude:
                  type: number
                  minimum: -90
                  maximum: 90
                  description: Latitude of the search origin
                  example: 37.7749
                longitude:
                  type: number
                  minimum: -180
                  maximum: 180
                  description: Longitude of the search origin
                  example: -122.4194
      responses:
        200:
          description: Array of pins within 10 km, sorted nearest first
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Pin'
        400:
          description: Bad request - missing or invalid coordinates
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        401:
          description: Unauthorized - invalid or missing token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /api/pins/{id}:
    get:
      tags: [Pins]
      summary: Get a specific pin by ID
      description: Retrieves a single pin by its ID
      operationId: getPin
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
          description: The ID of the pin to retrieve
          example: 1
      responses:
        200:
          description: Successful response with pin data
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Pin'
        400:
          description: Bad request (invalid path parameter)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        401:
          description: Unauthorized - invalid or missing token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    
    put:
      tags: [Pins]
      summary: Update a specific pin by ID
      description: Updates title/message/image/address for a pin (owner only). Returns 404 for both missing pins and pins owned by another user to prevent ID enumeration.
      operationId: updatePin
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
          description: The ID of the pin to update

      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                title:
                  type: string
                  maxLength: 200
                  nullable: true
                description:
                  type: string
                  maxLength: 500
                  nullable: true
                image:
                  type: string
                  maxLength: 2000
                  nullable: true
                address:
                  type: string
                  maxLength: 200
                  nullable: true

      responses:
        200:
          description: Pin updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Pin'
        400:
          description: Bad request (invalid body or path parameter)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        401:
          description: Unauthorized - invalid or missing token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        403:
          description: Unauthorized - not the owner
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        404:
          description: Pin not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

    delete:
      tags: [Pins]
      summary: Delete a specific pin by ID
      description: Delete a single pin by its ID (owner only)
      operationId: deletePin
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
          description: The ID of the pin to delete
      responses:
        200:
          description: Successfully deleted pin
          content: {}
        400:
          description: Bad request (invalid path parameter)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        401:
          description: Unauthorized - invalid or missing token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        403:
          description: Unauthorized - not the owner
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        404:
          description: Pin not found
          content: {}
        500:
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
  /api/pins/{id}/similar:
    get:
      tags: [Pins]
      summary: Get similar pins
      description: Returns up to 10 pins that users who liked this pin also liked
      operationId: getSimilarPins
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        200:
          description: Similar pins
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Pin'
        404:
          description: Pin not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    Pin:
      type: object
      properties:
        id:
          type: integer
          description: Unique identifier for the pin
          example: 1
        creatorID:
          type: integer
          description: ID of the account that created the pin
          example: 1
        email:
          type: string
          format: email
          description: Email of the pin creator
          example: "user@example.com"
        latitude:
          type: number
          description: self explanatory
        longitude:
          type: number
          description: self explanatory
        title:
          type: string
          maxLength: 200
          description: Title of the pin
          example: "Golden Gate Overlook"
          nullable: true
        description:
          type: string
          maxLength: 500
          description: Description/message content of the pin
          example: "Welcome to Beacon!"
          nullable: true
        image:
          type: string
          maxLength: 2000
          description: URL of the image
          example: "https://images.unsplash.com/photo-1516192518150-0d8fee5425e3"
          nullable: true
        address:
          type: string
          maxLength: 200
          description: Address of the pin location
          example: "123 Main St, San Francisco, CA"
          nullable: true
        tags:
          type: string
          maxLength: 500
          description: Tags stored as a JSON-encoded string array
          example: '["scenic","viewpoint","nature"]'
          nullable: true
        likes:
          type: integer
          description: Base like count stored on the pin
          example: 0
          nullable: true
        createdAt:
          type: string
          description: Timestamp when the pin was created
          example: "2026-01-17 12:00:00"
          nullable: true
        userStatus:
          type: string
          enum: [visited, wishlist]
          nullable: true
          description: |
            The requesting user's status for this pin, if any. Only populated on
            authenticated endpoints.
        trendingScore:
          type: number
          nullable: true
          description: |
            Computed trending score `likes + 3 * max(0, 1 - ageDays / days)`.
            Only present on `/api/pins/trending`.
...
    Error:
      type: object
      properties:
        message:
          type: string
          description: Human-readable error message
          example: "Invalid token"
        error:
          oneOf:
            - type: string
            - type: object
          description: Error identifier/message (some endpoints return 'error' instead of 'message')
          example: "Failed to share itinerary"
        errors:
          type: array
          description: Detailed validation errors (OpenAPI validator)
          items:
            type: object
        details:
          type: string
          description: Additional error details
          example: "FOREIGN KEY constraint failed"
...